devicestatus/requirements.txt
requests==2.22.0
wiotp-sdk
numpy
pandas
psycopg2

